requirements.txt
fastapi==0.103.2
uvicorn[standard]
Pillow==10.0.1
pillow-avif-plugin
qrcode==7.4.2
torch
diffusers["torch"] 
transformers
accelerate
xformers
python-multipart
pyzbar
sqlalchemy
# pip install torch==2.0.1+cu118 --index-url https://download.pytorch.org/whl/cu118